Refactor all the filez
