+++ ballet/validation/__init__.py
@@ -33,69 +33,6 @@ class BalletTestTypes:
     FEATURE_PRUNING_EVALUATION = 'feature_pruning_evaluation'
 
 
-def get_proposed_feature(project):
-    """Get the proposed feature
-
-    The path of the proposed feature is determined by diffing the project
-    against a comparison branch, such as master. The feature is then imported
-    from that path and returned.
-
-    Args:
-        project (ballet.project.Project): project info
-
-    Raises:
-        ballet.exc.BalletError: more than one feature collected
-    """
-    change_collector = ChangeCollector(project)
-    collected_changes = change_collector.collect_changes()
-    try:
-        new_feature_info = one_or_raise(collected_changes.new_feature_info)
-        importer, _, _ = new_feature_info
-    except ValueError:
-        raise BalletError('Too many features collected')
-    module = importer()
-    feature = _get_contrib_feature_from_module(module)
-    return feature
-
-
-def get_accepted_features(features, proposed_feature):
-    """Deselect candidate features from list of all features
-
-    Args:
-        features (List[Feature]): collection of all features in the ballet
-            project: both accepted features and candidate ones that have not
-            been accepted
-        proposed_feature (Feature): candidate feature that has not been
-            accepted
-
-    Returns:
-        List[Feature]: list of features with the proposed feature not in it.
-
-    Raises:
-        ballet.exc.BalletError: Could not deselect exactly the proposed
-            feature.
-    """
-    def eq(feature):
-        """Features are equal if they have the same source
-
-        At least in this implementation...
-        """
-        return feature.source == proposed_feature.source
-
-    # deselect features that match the proposed feature
-    result = lfilter(complement(eq), features)
-
-    if len(features) - len(result) == 1:
-        return result
-    elif len(result) == len(features):
-        raise BalletError(
-            'Did not find match for proposed feature within \'contrib\'')
-    else:
-        raise BalletError(
-            'Unexpected condition (n_features={}, n_result={})'
-            .format(len(features), len(result)))
-
-
 def detect_target_type():
     if TEST_TYPE_ENV_VAR in os.environ:
         return os.environ[TEST_TYPE_ENV_VAR]
