--- ballet/validation/feature_evaluation.py
@@ -1,14 +1,0 @@
-from ballet.validation.base import (
-    FeatureAcceptanceEvaluator, FeaturePruningEvaluator)
-
-
-class NoOpAcceptanceEvaluator(FeatureAcceptanceEvaluator):
-
-    def judge(self, feature):
-        return True
-
-
-class NoOpPruningEvaluator(FeaturePruningEvaluator):
-
-    def prune(self, k=4):
-        return []
