+++ ballet/validation/feature_pruning/validator.py
@@ -1,0 +1,4 @@
+class NoOpPruningEvaluator(FeaturePruningEvaluator):
+
+    def prune(self, k=4):
+        return []
